Move components to sub `DetectionStatus` folder
