+++ WEB/frontend/src/components/DetectionStatus/AppliedFilter.js
