+++ WEB/frontend/src/components/DetectionStatus/FilterBar.js
