+++ WEB/frontend/src/components/DetectionStatus/FilterCheckbox.js
