+++ WEB/frontend/src/components/DetectionStatus/FilterTable.js
