+++ WEB/frontend/src/components/DetectionStatus/ScrapButton.js
